refactor(backend): split to external graders
